Refactor xrons data model
Gets rid of the xrons table by 1) reassociating c14s and typos (evt.
dendros) to samples directly and 2) reassociating references with c14s
and typos (evt. dendros) through a polymorphic join table "citations".

replaced_by, replacement_comment and measurement_state are lost for now,
but this functionality is currently unused. Should be reimplemented in a
module or similar.