Implement single search API. Closes #140
With https://github.com/Casecommons/pg_search. Switching to this for
autocomplete with tom-select fixes #141. For now I've just implemented
it for the models/fields we filter in the browser view (i.e. site.name,
site_type.name, material.name, taxon.name, and c14.lab_identifier) but
this could be generalised to a site-wide search/filtering API.

Rewrote some (very) outdated JSON templates along the way. There are
ore of these.

Need to test compatibility with V1 API (see #146) and consider whether
it makes sense to maintain a separate search action for this, or just
fold it into index.